rename association to frequent_patterns and prepare v0.6 release (#163)
