+++ mlxtend/__init__.py
@@ -4,4 +4,4 @@
 #
 # License: BSD 3 clause
 
-__version__ = '0.6.0.dev0'
+__version__ = '0.6.0'
