+++ mlxtend/frequent_patterns/apriori.py
