+++ mlxtend/frequent_patterns/tests/test_apriori.py
@@ -5,7 +5,7 @@
 # License: BSD 3 clause
 
 import numpy as np
-from mlxtend.association import apriori
+from mlxtend.frequent_patterns import apriori
 from numpy.testing import assert_array_equal
 import pandas as pd
