+++ mlxtend/regressor/tests/test_stacking_regression.py
@@ -30,9 +30,9 @@ def test_different_models():
     stregr = StackingRegressor(regressors=[svr_lin, lr, ridge],
                                meta_regressor=svr_rbf)
     stregr.fit(X1, y).predict(X1)
-    mse = 0.214
+    mse = 0.21
     got = np.mean((stregr.predict(X1) - y) ** 2)
-    assert round(got, 3) == mse
+    assert round(got, 2) == mse
 
 
 def test_multivariate():
@@ -43,9 +43,9 @@ def test_multivariate():
     stregr = StackingRegressor(regressors=[svr_lin, lr, ridge],
                                meta_regressor=svr_rbf)
     stregr.fit(X2, y).predict(X2)
-    mse = 0.218
+    mse = 0.22
     got = np.mean((stregr.predict(X2) - y) ** 2)
-    assert round(got, 3) == mse
+    assert round(got, 2) == mse
 
 
 def test_multivariate_class():
@@ -55,9 +55,11 @@ def test_multivariate_class():
     stregr = StackingRegressor(regressors=[lr, ridge],
                                meta_regressor=meta)
     stregr.fit(X2, y2).predict(X2)
-    mse = 0.122
-    got = np.mean((stregr.predict(X2) - y2) ** 2)
-    assert round(got, 3) == mse
+    mse = 0.12
+    got = np.mean((stregr.predict(X2) - y2) ** 2.)
+    # there seems to be an issue with the following test on Windows
+    # sometimes via Appveyor
+    assert round(got, 2) == mse, got
 
 
 def test_gridsearch():
@@ -78,8 +80,8 @@ def test_gridsearch():
                         refit=True,
                         verbose=0)
     grid = grid.fit(X1, y)
-    best = 0.096
-    got = round(grid.best_score_, 3)
+    best = 0.1
+    got = round(grid.best_score_, 2)
     assert best == got
 
 
@@ -101,8 +103,8 @@ def test_gridsearch_numerate_regr():
                         refit=True,
                         verbose=0)
     grid = grid.fit(X1, y)
-    best = 0.096
-    got = round(grid.best_score_, 3)
+    best = 0.1
+    got = round(grid.best_score_, 2)
     assert best == got
 
 
@@ -126,8 +128,8 @@ def test_get_intercept():
                                meta_regressor=ridge)
     stregr.fit(X1, y)
     got = stregr.intercept_
-    expect = 0.024
-    assert round(got, 3) == expect
+    expect = 0.02
+    assert round(got, 2) == expect
 
 
 @raises(ValueError)
